genshin/requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
aiohttp==3.9.1
asyncio
genshin==1.7.3
cryptography==41.0.7